Add eslint and prettier configs
