+++ frontend/.eslintrc
@@ -1,44 +1,75 @@
 {
-  "extends": [
-    "plugin:react/recommended",
-    "airbnb"
-  ],
+  "extends": ["plugin:react/recommended", "prettier"],
+  "plugins": ["react", "prettier"],
+  "ignorePatterns": ["**/*.json"],
   "env": {
     "browser": true
   },
   "rules": {
-    "quotes": [
-      "error",
-      "single"
+    "prettier/prettier": [
+      "warn",
+      {
+        "tabWidth": 2,
+        "semi": true,
+        "singleQuote": false,
+        "endOfLine": "auto"
+      }
     ],
-    "comma-dangle": 0,
-    "consistent-return": 0,
-    "function-paren-newline": [
-      0,
-      "never"
+    "no-console": [
+      "warn",
+      {
+        "allow": ["warn", "error"]
+      }
     ],
-    "implicit-arrow-linebreak": [
-      "off"
+    "no-duplicate-imports": "warn",
+    "react/react-in-jsx-scope": "off",
+    "react/prop-types": 0,
+    "react/jsx-props-no-spreading": "off",
+    "react/jsx-curly-brace-presence": [
+      "warn",
+      {
+        "props": "never"
+      }
     ],
-    "no-param-reassign": 0,
-    "no-underscore-dangle": 0,
-    "no-shadow": 0,
-    "no-console": 0,
-    "no-plusplus": 0,
-    "no-unused-expressions": 0,
+    "react/jsx-boolean-value": 1,
+    "react/jsx-no-useless-fragment": [
+      "warn",
+      {
+        "allowExpressions": true
+      }
+    ],
+    "react/jsx-filename-extension": [
+      1,
+      {
+        "extensions": [".js", ".jsx", "ts", "tsx"]
+      }
+    ],
+    "eqeqeq": ["error", "smart"],
+    "import/extensions": "off",
+    "no-return-await": "warn",
     "no-unused-vars": [
       "error",
       {
         "argsIgnorePattern": "next"
       }
     ],
+    "no-var": "error",
+    "one-var": ["error", "never"],
+    "prefer-const": "warn",
+    "quotes": ["warn", "double"],
+    "comma-dangle": 0,
+    "consistent-return": 0,
+    "function-paren-newline": [0, "never"],
+    "implicit-arrow-linebreak": ["off"],
+    "no-param-reassign": 0,
+    "no-underscore-dangle": 0,
+    "no-shadow": 0,
+    "no-plusplus": 0,
+    "no-unused-expressions": 0,
     "import/prefer-default-export": "warn",
     "import/no-unresolved": "off",
     "linebreak-style": 0
   },
-  "plugins": [
-    "react"
-  ],
   "parserOptions": {
     "ecmaFeatures": {
       "jsx": true
