+++ frontend/.prettierignore
@@ -1,0 +1,13 @@
+# Ignore artifacts:
+build
+coverage
+public/resources
+.prettierignore
+package-lock.json
+.env*
+.gitignore
+*.ico
+
+# IDE
+.vscode
+.idea
