+++ frontend/.prettierrc.json
@@ -1,0 +1,6 @@
+{
+  "tabWidth": 2,
+  "semi": true,
+  "singleQuote": false,
+  "endOfLine": "auto"
+}
