+++ frontend/README.md
@@ -69,12 +69,11 @@ This section has moved here: [https://facebook.github.io/create-react-app/docs/d
 
 This section has moved here: [https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify](https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify)
 
-
 ## Working with MUX Spaces
 
 Follow the MUX docs [here](https://docs.mux.com/guides/video/build-real-time-video-experiences).
 
->If you want to quickly generate a one-off JWT for testing, you might want to try the [mux-cli](https://github.com/muxinc/cli), and use the `spaces:sign` [command](https://github.com/muxinc/cli#mux-spacessign-space-id) to quickly get a JWT.
+> If you want to quickly generate a one-off JWT for testing, you might want to try the [mux-cli](https://github.com/muxinc/cli), and use the `spaces:sign` [command](https://github.com/muxinc/cli#mux-spacessign-space-id) to quickly get a JWT.
 
 Set the following env variable to the signed JWT:
